+++ custom_components/freshharvest/__init__.py
@@ -10,7 +10,7 @@ from homeassistant.helpers.aiohttp_client import async_create_clientsession
 from .api import FreshHarvestClient
 from .coordinator import FreshHarvestCoordinator
 
-PLATFORMS: list[Platform] = [Platform.SENSOR]
+PLATFORMS: list[Platform] = [Platform.BINARY_SENSOR, Platform.SENSOR]
 
 type FreshHarvestConfigEntry = ConfigEntry[FreshHarvestCoordinator]